backend/requirements.txt
# ── Web Framework ──────────────────────────────────────────
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12          # file uploads

# ── Config & Validation ────────────────────────────────────
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1

# ── LLM Providers ──────────────────────────────────────────
openai==1.59.4
tiktoken==0.8.0                   # token counting for OpenAI

# ── HuggingFace Ecosystem ──────────────────────────────────
transformers==4.47.1
accelerate==1.2.1
datasets==3.2.0
tokenizers==0.21.0
sentencepiece==0.2.0

# ── LoRA / PEFT Fine-tuning ────────────────────────────────
peft==0.14.0
bitsandbytes==0.45.0              # 4-bit/8-bit quantization
trl==0.13.0                       # SFT trainer

# ── Embeddings & Vector Store ──────────────────────────────
faiss-cpu==1.9.0.post1
sentence-transformers==3.3.1
numpy==1.26.4

# ── Document Ingestion ─────────────────────────────────────
pypdf==5.1.0
python-docx==1.1.2
markdown==3.7

# ── Caching & Persistence ──────────────────────────────────
diskcache==5.6.3

# ── Observability ──────────────────────────────────────────
structlog==24.4.0
rich==13.9.4                      # pretty dev logging

# ── HTTP Client ────────────────────────────────────────────
httpx==0.28.1

# ── Rate Limiting ──────────────────────────────────────────
slowapi==0.1.9

# ── Testing ────────────────────────────────────────────────
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1                     # async test client
pytest-cov==6.0.0
